refactor(S2): useGamePlay 커스텀 훅 분리
