+++ apps/game-builder/src/components/game-play/play/GamePlay.tsx
@@ -15,7 +15,7 @@ export default function GamePlay({ playId, gameId, gameIntro }: GamePlayProps) {
   const pageId = gameIntro.play?.page?.id;
 
   if (!pageId) {
-    return router.push(`/game-play/start?gameId=${gameId}`);
+    router.push(`/game-play/start?gameId=${gameId}`);
   }
 
   return (
